making edits

+++ simulator/resolvers/current_education_level_by_country_resolver.py
@@ -8,6 +8,6 @@ class CurrentEducationLevelByCountryResolver(ResolverBase[dict[str, float]]):
     dependencies = [VariableEnum.current_year, VariableEnum.projections_education_level_by_country]
 
     @classmethod
-    def resolve(cls, current_values: dict[VariableEnum, Any]) -> dict[str, int]:
+    def resolve(cls, current_values: dict[VariableEnum, Any]) -> dict[str, float]:
         projected_education_level_by_country_for_year = current_values[VariableEnum.projections_education_level_by_country][current_values[VariableEnum.current_year]]
         return {**projected_education_level_by_country_for_year}
